Update nix environment
This in turn started breaking `cargo fmt`:

```
$ cargo fmt
Warning: can't set `imports_granularity = Crate`, unstable features are only available in nightly channel.
```

It turns out that `imports_granularity` is not stabilized and somehow
this was flying under the radar before.

So just nuke it all and use the default of `Preserve`.